requirements.txt
absl-py==2.3.1
ale-py==0.10.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733175639022/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///D:/bld/cffi_1723018491042/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1722923915790/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
gymnasium==1.0.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
huggingface-hub==0.33.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1683289169355/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.6
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver==1.4.7
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mediapy==1.2.4
-e git+https://github.com/metadriverse/metadrive.git@85e5dadc6c7436d324348f6e3d8f8e680c06b4db#egg=metadrive_simulator
mpmath==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
numpy @ file:///D:/bld/numpy_1687808486521/work
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
Panda3D==1.10.15
panda3d-gltf==0.14
panda3d-simplepbr==0.12.0
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///D:/bld/pillow_1675487410089/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
progressbar==2.5
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
protobuf==5.29.5
psutil @ file:///D:/bld/psutil_1719274704414/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pygame==2.6.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1724399214825/work
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
shapely==2.0.7
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stable_baselines3==2.4.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tornado @ file:///D:/bld/tornado_1717722870081/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.46.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.6
win_inet_pton @ file:///D:/bld/win_inet_pton_1727796272493/work
yapf==0.43.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1731262100163/work
zstandard==0.19.0
